refactor(provider): split openai chat backend
